refactor to share class fetching behavior
