+++ src/app/api/classes/route.ts
@@ -1,12 +1,12 @@
 import {NextResponse} from 'next/server';
-import {classes} from './constants';
+import {getAllClasses} from '@/lib/classes';
 
 export async function GET() {
 	// in the future grab this data from contentful or other headless cms with try/catch block
 	return NextResponse.json(
 		{
 			message: 'OK',
-			classes,
+			classes: getAllClasses(),
 		},
 		{status: 200},
 	);
