+++ src/lib/classes/constants.ts
